+++ controller.js
@@ -38,53 +38,8 @@ module.exports = {
         } catch(e) {
             onError(e, res);
         }
-    },
-    generate: function(req, res, next) {
-        var start = req.param("start");
-        var end = req.param("end");
-        start = moment(start);
-        end = moment(end);
-        var date = moment(start);
-
-        function callback() {
-            date = moment(date).add(1, 'days');
-            if (moment(date).isAfter(end) || moment(date).isAfter(moment())) {
-                console.log("Complete", start.format("MM/DD/YYYY"), end.format("MM/DD/YYYY"));
-                res.send("Generated data for all games between " + moment(start).format("MM/DD/YYYY") + ' and ' + moment(end).format('MM/DD/YYYY'));
-            } else {
-                getGamesDataForDate(date)
-            }
-        }
-
-        getGamesDataForDate(date, callback, function(e) {console.log(e); callback();});
     }
 };
-
-function getGamesDataForDate(date, callback, error) {
-    getGamesForDate(date.toDate(), function(games) {
-        if( games && games.length ) {
-            var promisesForGames = [];
-            _.each(games, function(game) {
-
-                var homeTeam = _.findWhere(nba.teamsInfo, {teamId: game.homeTeamId});
-                var visitorTeam = _.findWhere(nba.teamsInfo, {teamId: game.visitorTeamId});
-
-                promisesForGames.push(service.getGameStatsFromApi(game, homeTeam, date));
-                promisesForGames.push(service.getGameStatsFromApi(game, visitorTeam, date));
-            });
-            Promise.all(promisesForGames).then(callback, error);
-        } else {
-            callback();
-        }
-    })
-}
-
-function getGamesForDate(date, callback, error) {
-    nba.api.scoreboard({ GameDate: moment(date).format('MM/DD/YYYY') }).then(function(resp) {
-        callback(resp.gameHeader);
-    }).catch(error);
-}
-
 
 function getLastGameForTeam(teamId, date, callback, error) {
     if (!date) {
@@ -96,7 +51,7 @@ function getLastGameForTeam(teamId, date, callback, error) {
             return game.homeTeamId == teamId || game.visitorTeamId == teamId
         });
         if ( !teamGame || teamGame.gameStatusId != GAME_STATUS_FINAL ) {
-            date = moment(date).add(1, 'days').toDate();
+            date = moment(date).subtract(1, 'days').toDate();
             getLastGameForTeam(teamId, date, callback);
         } else {
             global.date = date;
